+++ src/test/java/org/apache/aurora/scheduler/reconciliation/TaskReconcilerTest.java
@@ -17,6 +17,7 @@ import java.util.List;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicLong;
 
+import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
@@ -97,20 +98,20 @@ public class TaskReconcilerTest extends EasyMockTest {
     storageUtil.expectTaskFetch(
         Query.unscoped().byStatus(Tasks.SLAVE_ASSIGNED_STATES),
         task1,
-        task2).times(5);
+        task2).times(7);
 
     List<List<Protos.TaskStatus>> batches = Lists.partition(ImmutableList.of(
         TASK_TO_PROTO.apply(task1),
         TASK_TO_PROTO.apply(task2)), BATCH_SIZE);
 
     driver.reconcileTasks(batches.get(0));
-    expectLastCall().times(5);
+    expectLastCall().times(7);
 
     driver.reconcileTasks(batches.get(1));
-    expectLastCall().times(5);
+    expectLastCall().times(7);
 
     driver.reconcileTasks(EasyMock.anyObject());
-    expectLastCall().times(2);
+    expectLastCall().times(3);
 
     control.replay();
 
@@ -138,6 +139,15 @@ public class TaskReconcilerTest extends EasyMockTest {
     clock.advance(IMPLICT_SCHEDULE);
     assertEquals(5L, explicitRuns.get());
     assertEquals(2L, implicitRuns.get());
+
+    reconciler.triggerExplicitReconciliation(Optional.of(BATCH_SIZE));
+    assertEquals(6L, explicitRuns.get());
+    reconciler.triggerImplicitReconciliation();
+    assertEquals(3L, implicitRuns.get());
+
+    reconciler.triggerExplicitReconciliation(Optional.absent());
+    assertEquals(7L, explicitRuns.get());
+    assertEquals(3L, implicitRuns.get());
   }
 
   @Test(expected = IllegalArgumentException.class)