+++ src/test/python/apache/aurora/client/api/test_scheduler_client.py
@@ -34,6 +34,7 @@ import gen.apache.aurora.api.AuroraAdmin as AuroraAdmin
 import gen.apache.aurora.api.AuroraSchedulerManager as AuroraSchedulerManager
 from gen.apache.aurora.api.constants import BYPASS_LEADER_REDIRECT_HEADER_NAME
 from gen.apache.aurora.api.ttypes import (
+    ExplicitReconciliationSettings,
     Hosts,
     JobConfiguration,
     JobKey,
@@ -286,6 +287,18 @@ class TestSchedulerProxyAdminInjection(TestSchedulerProxyInjection):
         IsA(RewriteConfigsRequest)).AndReturn(DEFAULT_RESPONSE)
     self.mox.ReplayAll()
     self.make_scheduler_proxy().rewriteConfigs(RewriteConfigsRequest())
+
+  def test_triggerExplicitTaskReconciliation(self):
+    self.mock_thrift_client.triggerExplicitTaskReconciliation(
+      IsA(ExplicitReconciliationSettings)).AndReturn(DEFAULT_RESPONSE)
+    self.mox.ReplayAll()
+    self.make_scheduler_proxy().triggerExplicitTaskReconciliation(
+      ExplicitReconciliationSettings(batchSize=None))
+
+  def test_triggerImplicitTaskReconciliation(self):
+    self.mock_thrift_client.triggerImplicitTaskReconciliation().AndReturn(DEFAULT_RESPONSE)
+    self.mox.ReplayAll()
+    self.make_scheduler_proxy().triggerImplicitTaskReconciliation()
 
 
 def mock_auth():